Refactor irc decorators into seperate module
