+++ app/handlers/irc/authentication.py
@@ -1,21 +1,9 @@
 
 from config import OSU_IRC_ENABLED
 from typing import List, Optional, Callable
+from app.handlers.irc.decorators import *
 from app.clients.irc import IrcClient
 from app import session
-
-def register(command: str) -> Callable:
-    def wrapper(func) -> Callable:
-        session.irc_handlers[command] = func
-        return func
-    return wrapper
-
-def ensure_unauthenticated(func: Callable) -> Callable:
-    def wrapper(client: IrcClient, *args, **kwargs) -> None:
-        if client.logged_in:
-            return
-        return func(client, *args, **kwargs)
-    return wrapper
 
 @register("USER")
 @ensure_unauthenticated
