+++ app/handlers/irc/chat.py
@@ -1,24 +1,12 @@
 
 from typing import List, Optional, Callable
 from twisted.words.protocols import irc
+from app.handlers.irc.decorators import *
 from app.common.database import messages
 from app.clients.irc import IrcClient
 from app import session
 
 import time
-
-def register(command: str) -> Callable:
-    def wrapper(func) -> Callable:
-        session.irc_handlers[command] = func
-        return func
-    return wrapper
-
-def ensure_authenticated(func: Callable) -> Callable:
-    def wrapper(client: IrcClient, *args, **kwargs) -> None:
-        if not client.logged_in:
-            return
-        return func(client, *args, **kwargs)
-    return wrapper
 
 @register("LIST")
 @ensure_authenticated
