+++ app/handlers/irc/irc.py
@@ -1,14 +1,6 @@
 
-from typing import List, Optional, Callable
-from twisted.words.protocols import irc
+from app.handlers.irc.decorators import *
 from app.clients.irc import IrcClient
-from app import session
-
-def register(command: str) -> Callable:
-    def wrapper(func) -> Callable:
-        session.irc_handlers[command] = func
-        return func
-    return wrapper
 
 @register("PING")
 def handle_ping_command(
