+++ app/handlers/irc/players.py
@@ -1,22 +1,7 @@
 
-
-from typing import List, Optional, Callable
-from twisted.words.protocols import irc
+from app.handlers.irc.decorators import *
 from app.clients.irc import IrcClient
 from app import session
-
-def register(command: str) -> Callable:
-    def wrapper(func) -> Callable:
-        session.irc_handlers[command] = func
-        return func
-    return wrapper
-
-def ensure_authenticated(func: Callable) -> Callable:
-    def wrapper(client: IrcClient, *args, **kwargs) -> None:
-        if not client.logged_in:
-            return
-        return func(client, *args, **kwargs)
-    return wrapper
 
 @register("NAMES")
 @ensure_authenticated
